Cleanup BaseApp and StatePanelView
For #124
